requirements.txt
PySide6==6.6.1
qasync==0.25.0
python-binance==1.0.19
aiohttp==3.9.3
pandas==2.1.3
numpy==1.26.0
pydantic==1.10.7
python-dotenv==0.19.2
transformers==4.40.1
ta==0.11.0
backtrader==1.9.78.123
textblob==0.17.1
requests==2.31.0
cachetools==5.3.2
asyncpraw==7.7.1